requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

bertopic==0.17.3
deepl==1.22.0
emoji==2.14.1
matplotlib==3.9.4
nltk==3.9.1
ollama==0.5.3
pandas==2.3.0
plotly==6.3.0
python-dateutil==2.9.0.post0
sentence-transformers==5.1.0
streamlit==1.49.1
Telethon==1.28.5
torch==2.7.1
tqdm==4.65.0
transformers==4.53.3
umap-learn==0.5.9.post2
